[Security Solution][Resolver] standardize resolver panel component naming (#74537)
